+++ test/phoenix_live_view/integrations/flash_test.exs
@@ -129,7 +129,7 @@ defmodule Phoenix.LiveView.FlashIntegrationTest do
     test "nested redirect with flash", %{conn: conn} do
       {:ok, flash_live, _} = live(conn, "/flash-root")
 
-      flash_child = get_live_child(flash_live, "flash-child")
+      flash_child = find_live_child(flash_live, "flash-child")
       render_click(flash_child, "redirect", %{"to" => "/flash-root?redirect", "info" => "ok!"})
       flash = assert_redirect(flash_child, "/flash-root?redirect")
       assert flash == %{"info" => "ok!"}
@@ -138,7 +138,7 @@ defmodule Phoenix.LiveView.FlashIntegrationTest do
     test "nested push_redirect with flash", %{conn: conn} do
       {:ok, flash_live, _} = live(conn, "/flash-root")
 
-      flash_child = get_live_child(flash_live, "flash-child")
+      flash_child = find_live_child(flash_live, "flash-child")
       render_click(flash_child, "push_redirect", %{"to" => "/flash-root?push", "info" => "ok!"})
       flash = assert_redirect(flash_child, "/flash-root?push")
       assert flash == %{"info" => "ok!"}
@@ -147,7 +147,7 @@ defmodule Phoenix.LiveView.FlashIntegrationTest do
     test "nested push_patch with flash", %{conn: conn} do
       {:ok, flash_live, _} = live(conn, "/flash-root")
 
-      flash_child = get_live_child(flash_live, "flash-child")
+      flash_child = find_live_child(flash_live, "flash-child")
       render_click(flash_child, "push_patch", %{"to" => "/flash-root?patch", "info" => "ok!"})
 
       result = render(flash_live)